Refactored AuditTrailComponent and plugged into view
