--- app/components/audit_trail_component.html.erb
@@ -1,26 +1,0 @@
-<% @application_form.audits.each do |audit| %>
-  <tr class="govuk-table__row gov uk-accordion__section">
-    <td class="govuk-table__cell"><%= audit.created_at.to_s(:db) %></td>
-    <td class="govuk-table__cell">
-      <details class="govuk-details govuk-!-margin-0" data-module="govuk-details">
-        <summary class="govuk-details__summary">
-          <%= audit_entry_label(audit) %>
-        </summary>
-        <div class="govuk-details__text">
-          <dl class="govuk-summary-list--no-border">
-            <% audit.audited_changes.select { |_, values| audit_value_present?(values) }.each do |attr, values| %>
-              <div class="govuk-summary-list__row">
-                <dt class="govuk-summary-list__key">
-                  <%= attr %>
-                </dt>
-                <dd class="govuk-summary-list__value">
-                  <%= format_audit_value(values) %>
-                </dd>
-              </div>
-            <% end %>
-          </dl>
-        </div>
-      </details>
-    </td>
-  </tr>
-<% end %>
